chore: minor directory change
