+++ lib/command.txt
