+++ lib/main.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:on_time_front/screens/schedule_list.dart';
-import 'package:on_time_front/screens/server_test/schedule_list_test.dart';
+import 'package:on_time_front/presentation/preparation/screens/schedule_list.dart';
+import 'package:on_time_front/presentation/preparation/screens/server_test/schedule_list_test.dart';
 import 'package:on_time_front/shared/theme/theme.dart';
 import 'package:on_time_front/widgets/login_buttons/login_test_screen.dart';
