+++ lib/presentation/preparation/component/preparation_timer/preparation_step.dart
