+++ lib/presentation/preparation/screens/schedule_list.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:on_time_front/core/utils/data_loader.dart';
-import 'package:on_time_front/screens/schedule_start.dart';
+import 'package:on_time_front/presentation/preparation/screens/schedule_start.dart';
 
 class ScheduleList extends StatefulWidget {
   const ScheduleList({super.key});
