+++ lib/presentation/preparation/screens/schedule_start.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
-import 'package:on_time_front/screens/server_test/alarm_screen_test.dart';
-import 'package:on_time_front/widgets/preparation_timer/button.dart';
+import 'package:on_time_front/presentation/preparation/screens/server_test/alarm_screen_test.dart';
+import 'package:on_time_front/presentation/preparation/component/preparation_timer/button.dart';
 
-class ScheduleStartTest extends StatelessWidget {
+class ScheduleStart extends StatelessWidget {
   final Map<String, dynamic> schedule;
 
-  const ScheduleStartTest({
+  const ScheduleStart({
     super.key,
     required this.schedule,
   });
@@ -34,7 +34,7 @@ class ScheduleStartTest extends StatelessWidget {
                   ),
                   const SizedBox(height: 8),
                   Text(
-                    schedule['place']['placeName'],
+                    schedule['placeName'],
                     style: const TextStyle(
                       fontSize: 25,
                       fontWeight: FontWeight.bold,
@@ -70,6 +70,10 @@ class ScheduleStartTest extends StatelessWidget {
                 Navigator.push(
                   context,
                   MaterialPageRoute(
+                    // builder: (context) => AlarmScreenNew(
+                    //   // schedule 데이터를 AlarmScreen으로 전달
+                    //   schedule: schedule,
+                    // ),
                     builder: (context) => AlarmScreenTest(
                       // schedule 데이터를 AlarmScreen으로 전달
                       schedule: schedule,
